+++ pom.xml
@@ -1,7 +1,7 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
   xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
-  <groupId>org.esmerilprogramming</groupId>
+  <groupId>org.jboss.aesh</groupId>
   <artifactId>cloverxell</artifactId>
   <version>0.56</version>
   <name>cloverxell</name>
@@ -49,7 +49,7 @@
             <manifest>
               <addClasspath>true</addClasspath>
               <classpathPrefix>lib/</classpathPrefix>
-              <mainClass>org.esmerilprogramming.cloverxell.AppMain</mainClass>
+              <mainClass>org.jboss.aesh.cloverxell.AppMain</mainClass>
             </manifest>
             <manifestEntries>
               <Class-Path>.</Class-Path>
